Added course owner relation + forgetpassword

+++ controller/courses.js
@@ -24,7 +24,7 @@ module.exports.getCourses = asyncHandler(async (req, res, next) => {
 // @route       GET /api/v1/courses/:courseID
 // @access      Public
 module.exports.getCourse = asyncHandler(async (req, res, next) => {
-  let course = await Course.findById(req.params.id).populate({
+  const course = await Course.findById(req.params.id).populate({
     path: "bootcamp",
     select: "name description",
   });
@@ -40,8 +40,9 @@ module.exports.getCourse = asyncHandler(async (req, res, next) => {
 // @route       POST /api/v1/bootcamp/:bootcampId/courses
 // @access      Private
 module.exports.addCourse = asyncHandler(async (req, res, next) => {
-  let bootcamp = await Bootcamp.findById(req.params.bootcampId);
+  const bootcamp = await Bootcamp.findById(req.params.bootcampId);
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = bootcamp.user;
 
   if (!bootcamp) {
     return next(
@@ -49,8 +50,14 @@ module.exports.addCourse = asyncHandler(async (req, res, next) => {
     );
   }
 
-  const course = await Course.create(req.body);
+  //making sure user id owner of bootcamp
+  if (req.user.id !== bootcamp.user.toString() && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse("Not authorized to add course to this bootcamp", 401)
+    );
+  }
 
+  const course = await Course.create(req.body);
   res.status(200).json({ success: true, count: course.length, Data: course });
 });
 
@@ -59,15 +66,22 @@ module.exports.addCourse = asyncHandler(async (req, res, next) => {
 // @access      Private
 module.exports.updateCourse = asyncHandler(async (req, res, next) => {
   let course = await Course.findById(req.params.id);
+  //checking if course exits
   if (!course) {
     return next(new ErrorResponse(`Course not found ${req.params.id}`, 404));
   }
-
+  //checking owner of course
+  if (req.user.id !== course.user.toString() && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse("Not authorized to Update course to this bootcamp", 401)
+    );
+  }
+  //updating course
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
     runValidators: true,
     new: true,
   });
-
+  //sending request
   res.status(200).json({ success: true, count: course.length, Data: course });
 });
 
@@ -75,9 +89,14 @@ module.exports.updateCourse = asyncHandler(async (req, res, next) => {
 // @route       Delete /api/v1/courses/:id
 // @access      Private
 module.exports.deleteCourse = asyncHandler(async (req, res, next) => {
-  let course = await Course.findById(req.params.id);
+  const course = await Course.findById(req.params.id);
   if (!course) {
     return next(new ErrorResponse(`Course not found ${req.params.id}`, 404));
+  }
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse("You are not authorized to delete this course!", 401)
+    );
   }
 
   await course.deleteOne();
